basic structure for calling diff migration functions
