--- slmigrate/migrate/capture.py
@@ -1,2 +1,0 @@
-def capture_migration(service):
-    print(service + " capture migration called")
